Create utils file Create shader classe create basic shader add wrapper in maths classes Correct window size and classe small restructuration of the engine

+++ FeldEngine-Core/src/Maths/Maths_function.h
@@ -8,7 +8,7 @@ namespace Fd {
 
 		float toRadians(float degrees)
 		{
-			return degrees * (M_PI / 180.0f);
+			return static_cast<float>(degrees * (M_PI / 180.0f));
 		}
 	}
 }
